Moving to beans.xml instead of extension loading beans.
I have not done any @Veto, I honestly don't think it buys us much right
now.

I have tested this with the wicket example on Glassfish 3.1 b36 and b37.
I imagine this should work in Glassfish 3.0.1, but I have not tested it
because the wicket example requires Weld 1.1, it's not portable :(